+++ app/admin/@toolbar/category/page.tsx
@@ -1,0 +1,3 @@
+import {Content} from './[...tabId]/content'
+const Page = async () => <Content />
+export default Page