+++ app/admin/category/layout.tsx
@@ -1,0 +1,6 @@
+import {PropsWithChildren} from 'react'
+
+const CategoryLayout = async ({children}: PropsWithChildren) => (
+  <div className='px-3'>{children}</div>
+)
+export default CategoryLayout