+++ app/admin/category/new-category.tsx
@@ -1,0 +1,12 @@
+'use client'
+
+import {Suspense} from 'react'
+import {CategoryForm} from './category-form'
+
+export const NewCategory = () => {
+  return (
+    <Suspense fallback={<div>Loading...</div>}>
+      <CategoryForm />
+    </Suspense>
+  )
+}