+++ app/admin/product/_badges/content.tsx
@@ -1,7 +1,0 @@
-'use client'
-
-import {BadgeList} from '../badges'
-
-export const Content = () => {
-  return <BadgeList />
-}