+++ app/admin/product/_new/page.tsx
@@ -1,3 +1,0 @@
-import {NewProduct} from '../new'
-const Page = async () => <NewProduct />
-export default Page